Rename test fixture helper package
